Include the contents of generated documents in the LSIF output
