+++ src/EditorFeatures/TestUtilities/Workspaces/TestWorkspace.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Formatting;
@@ -207,6 +208,9 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
 
         public new void OnParseOptionsChanged(ProjectId projectId, ParseOptions parseOptions)
             => base.OnParseOptionsChanged(projectId, parseOptions);
+
+        public new void OnAnalyzerReferenceAdded(ProjectId projectId, AnalyzerReference analyzerReference)
+            => base.OnAnalyzerReferenceAdded(projectId, analyzerReference);
 
         public void OnDocumentRemoved(DocumentId documentId, bool closeDocument = false)
         {
