+++ src/Features/Lsif/Generator/Generator.cs
@@ -13,6 +13,8 @@ using Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator.ResultSetTracki
 using Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator.Writing;
 using Methods = Microsoft.VisualStudio.LanguageServer.Protocol.Methods;
 using System.Collections.Concurrent;
+using System.Text;
+using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator
 {
@@ -86,7 +88,21 @@ namespace Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator
             var syntaxFactsService = languageServices.GetRequiredService<ISyntaxFactsService>();
             var semanticFactsService = languageServices.GetRequiredService<ISemanticFactsService>();
 
-            var documentVertex = new Graph.LsifDocument(new Uri(syntaxTree.FilePath), GetLanguageKind(semanticModel.Language), idFactory);
+            string? contentBase64Encoded = null;
+
+            // TODO: move to checking the enum member mentioned in https://github.com/dotnet/roslyn/issues/49326 when that
+            // is implemented. In the mean time, we'll use a heuristic of the path being a relative path as a way to indicate
+            // this is a source generated file.
+            if (!PathUtilities.IsAbsolute(syntaxTree.FilePath))
+            {
+                var text = semanticModel.SyntaxTree.GetText();
+
+                // We always use UTF-8 encoding when writing out file contents, as that's expected by LSIF implementations.
+                // TODO: when we move to .NET Core, is there a way to reduce allocatios here?
+                contentBase64Encoded = Convert.ToBase64String(Encoding.UTF8.GetBytes(text.ToString()));
+            }
+
+            var documentVertex = new Graph.LsifDocument(new Uri(syntaxTree.FilePath, UriKind.RelativeOrAbsolute), GetLanguageKind(semanticModel.Language), contentBase64Encoded, idFactory);
 
             lsifJsonWriter.Write(documentVertex);
             lsifJsonWriter.Write(new Event(Event.EventKind.Begin, documentVertex.GetId(), idFactory));
