+++ src/Features/Lsif/GeneratorTest/ProjectStructureTests.vb
@@ -2,8 +2,10 @@
 ' The .NET Foundation licenses this file to you under the MIT license.
 ' See the LICENSE file in the project root for more information.
 
+Imports System.Text
 Imports Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
 Imports Microsoft.CodeAnalysis.Test.Utilities
+Imports Roslyn.Test.Utilities
 
 Namespace Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator.UnitTests
     <UseExportProvider>
@@ -22,8 +24,40 @@ Namespace Microsoft.CodeAnalysis.LanguageServerIndexFormat.Generator.UnitTests
             Dim projectVertex = Assert.Single(lsif.Vertices.OfType(Of Graph.LsifProject))
             Dim documentVertices = lsif.GetLinkedVertices(Of Graph.LsifDocument)(projectVertex, "contains")
 
-            Assert.Single(documentVertices, Function(d) d.Uri.LocalPath = "Z:\A.cs")
-            Assert.Single(documentVertices, Function(d) d.Uri.LocalPath = "Z:\B.cs")
+            Dim documentA = Assert.Single(documentVertices, Function(d) d.Uri.LocalPath = "Z:\A.cs")
+            Dim documentB = Assert.Single(documentVertices, Function(d) d.Uri.LocalPath = "Z:\B.cs")
+
+            ' We don't include contents for normal files, just generated ones
+            Assert.Null(documentA.Contents)
+            Assert.Null(documentB.Contents)
+        End Function
+
+        <Fact>
+        Public Async Function SourceGeneratedDocumentsIncludeContent() As Task
+            Dim workspace = TestWorkspace.CreateWorkspace(
+                    <Workspace>
+                        <Project Language="C#" Name="TestProject" FilePath="Z:\TestProject.csproj" CommonReferences="true">
+                        </Project>
+                    </Workspace>)
+
+            workspace.OnAnalyzerReferenceAdded(workspace.CurrentSolution.ProjectIds.Single(),
+                                               New TestGeneratorReference(New TestSourceGenerator.HelloWorldGenerator()))
+
+            Dim lsif = Await TestLsifOutput.GenerateForWorkspaceAsync(workspace)
+
+            Dim projectVertex = Assert.Single(lsif.Vertices.OfType(Of Graph.LsifProject))
+            Dim generatedDocumentVertex = Assert.Single(lsif.GetLinkedVertices(Of Graph.LsifDocument)(projectVertex, "contains"))
+
+            ' Assert the contents were included and does match the tree
+            Dim contentBase64Encoded = generatedDocumentVertex.Contents
+            Assert.NotNull(contentBase64Encoded)
+
+            Dim contents = Encoding.UTF8.GetString(Convert.FromBase64String(contentBase64Encoded))
+
+            Dim compilation = Await workspace.CurrentSolution.Projects.Single().GetCompilationAsync()
+            Dim tree = Assert.Single(compilation.SyntaxTrees)
+
+            Assert.Equal(tree.GetText().ToString(), contents)
         End Function
     End Class
 End Namespace
